requirements.txt
pydantic
pyhumps
pyyaml
simpleobsws
wmi
